Display export format breaking change to the user
The serialized transaction format was changed here https://github.com/spesmilo/electrum/pull/4405

The change occurred in Electrum 3.2.0.

(
  before: https://github.com/spesmilo/electrum/blob/624fa4769d9f02fa71696b3d69030ae114ccf573/lib/transaction.py#L891
  after: https://github.com/spesmilo/electrum/blob/0119ab9ee1c4b07255442c7c91a7119ca989b64c/lib/transaction.py#L1013
)

I ran into this myself, here: https://github.com/spesmilo/electrum/issues/5924

Personally I believe that users should be notified of breaking changes.  Users should not have to make educated guesses, ask online, or be python programmers in order to be able to tell whats going on / why doesn't it work.

If you don't like this verbiage (for example, if you think that the user should be instructed to always use the exact same electrum version when importing / exporting)  

Then an alternative verbiage would be appropriate, such as "Always use the exact same Electrum version to import any exported transactions."

If that doesn't fit there the label can be moved. Or it could be  `'Electrum ' + current_electrum_version + ' required to import.'` Or Something. Point is that there should be some indication of this fact of life to the user.